pyproject.toml
[project]
name = "evolution"
version = "0.1.0"
description = "Multi-agent research & code evolution platform"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[project.scripts]
evolution = "evolution.cli.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
